refactor(playwright): decouple deploment code from visual-regression generator
