+++ packages/playwright/src/generators/visual-regression-deployment/schema.d.ts
@@ -1,0 +1,3 @@
+export interface VisualRegressionDeploymentGeneratorSchema {
+    visualRegression: string;
+}
