+++ packages/playwright/src/generators/visual-regression/files/applitools/applitools-eyes-grid.spec.ts__template__
