+++ packages/playwright/src/generators/visual-regression/files/native/playwright-visual-regression.spec.ts__template__
