+++ packages/playwright/src/generators/visual-regression/schema.d.ts
