+++ packages/playwright/src/generators/visual-regression/schema.json
