+++ packages/playwright/src/generators/visual-regression/utils/update-playwright-config.ts
