+++ packages/playwright/src/generators/visual-regression/utils/update-targets.ts
