+++ packages/playwright/src/index.ts
@@ -1,2 +1,5 @@
 export { default as initGenerator } from './generators/init/generator';
 export { default as initDeploymentGenerator } from './generators/init-deployment/generator';
+export { default as visualRegressionGenerator } from './generators/visual-regression/generator';
+export { default as visualRegressionDeploymentGenerator } from './generators/visual-regression-deployment/generator';
+export { default as accessibilityGenerator } from './generators/accessibility/generator';
