+++ packages/playwright/src/utils/types.ts
@@ -1,0 +1,4 @@
+export const visualRegressionTypes = {
+    NATIVE: 'native',
+    APPLITOOLS: 'applitools',
+};
